cinematch/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
torch==2.4.1
torchvision==0.19.1
ftfy==6.2.3
regex==2024.7.24
Pillow==10.4.0
chromadb==0.5.5
open-clip-torch==2.26.1
requests==2.32.3
yt-dlp>=2026.1.0
